requirements.txt
asgiref==3.7.2
dj-rest-auth==4.0.1
Django==4.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Faker==18.9.0
PyJWT==2.7.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
